@@ -16,16 +16,17 @@
     "generate": "vite build --ssrManifest --outDir dist/static && yarn build:server && node prerender",
     "debug": "node --inspect-brk server"
   },
+  "type": "module",
   "dependencies": {
-    "@live-change/cli": "0.7.33",
+    "@live-change/cli": "0.7.39",
     "@live-change/dao": "0.5.22",
     "@live-change/dao-vue3": "0.5.22",
     "@live-change/dao-websocket": "0.5.22",
-    "@live-change/framework": "0.7.33",
-    "@live-change/vue3-components": "0.2.29",
-    "@live-change/vue3-ssr": "0.2.29",
-    "@vitejs/plugin-vue": "^4.3.4",
-    "@vueuse/core": "^10.4.1",
+    "@live-change/framework": "0.7.39",
+    "@live-change/vue3-components": "0.2.34",
+    "@live-change/vue3-ssr": "0.2.34",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vueuse/core": "^10.7.2",
     "@vueuse/head": "^2.0.0",
     "codeceptjs-assert": "^0.0.5",
     "compression": "^1.7.4",
@@ -34,30 +35,30 @@
     "javascript-stringify": "^2.1.0",
     "primeflex": "^3.3.1",
     "primeicons": "^6.0.1",
-    "primevue": "^3.34.1",
+    "primevue": "^3.48.1",
     "prism-es6": "^1.2.0",
-    "prismjs": "^1.28.0",
+    "prismjs": "^1.29.0",
     "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-visualizer": "5.9.2",
-    "serialize-javascript": "^6.0.1",
+    "rollup-plugin-visualizer": "5.12.0",
+    "serialize-javascript": "^6.0.2",
     "serve-static": "^1.15.0",
     "v-shared-element": "3.1.1",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-vue-images": "^0.6.1",
-    "vue": "^3.2.47",
+    "vue": "^3.4.19",
     "vue-prism-editor": "2.0.0-alpha.2",
-    "vue-router": "^4.2.4",
-    "vue3-scroll-border": "0.1.5"
+    "vue-router": "^4.2.5",
+    "vue3-scroll-border": "0.1.6"
   },
   "devDependencies": {
-    "@live-change/codeceptjs-helper": "0.7.33",
+    "@live-change/codeceptjs-helper": "0.7.39",
     "@wdio/selenium-standalone-service": "^8.15.0",
-    "codeceptjs": "^3.5.4",
-    "generate-password": "1.7.0",
-    "playwright": "^1.37.1",
+    "codeceptjs": "^3.5.12",
+    "generate-password": "1.7.1",
+    "playwright": "^1.41.2",
     "random-profile-generator": "^2.3.0",
     "txtgen": "^3.0.6",
-    "webdriverio": "^8.16.6"
+    "webdriverio": "^8.31.1"
   },
   "author": "",
   "license": "ISC",
