@@ -1,4 +1,5 @@
-const app = require('@live-change/framework').app()
+import App from '@live-change/framework'
+const app = App.app()
 
 module.exports = async function(services) {
 
