@@ -1,10 +1,18 @@
-const test = require('tape')
-const levelup = require('levelup')
-const leveldown = require('leveldown')
-const encoding = require('encoding-down')
-const rimraf = require("rimraf")
+import test from 'tape'
+import levelup from 'levelup'
+import leveldown from 'leveldown'
+import encoding from 'encoding-down'
+import { rimraf } from 'rimraf'
 
-const Store = require('../lib/Store.js')
+import Store from '../lib/Store.js'
+
+import PQueue from 'p-queue'
+const testQueue = new PQueue({ concurrency: 1 })
+function queued(fn) {
+  return async t => {
+    return await testQueue.add(() => fn(t))
+  }
+}
 
 const dbPath = `./test.ro.db`
 rimraf.sync(dbPath)
@@ -51,39 +59,40 @@ test("store range observable", t => {
     let values = await getNextValue()
     t.deepEqual(values, [ { v: 1, id: 'a' }, { v: 3, id: 'c' } ], 'range value' )
 
-    t.test("remove object 'a' from observed range", async t => {
+    t.test("remove object 'a' from observed range", queued(async t => {
       t.plan(1)
       await store.delete('a')
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 3, id: 'c' } ], 'range value' )
-    })
+    }))
 
-    t.test("add object 'a' to observed range", async t => {
+    t.test("add object 'a' to observed range", queued(async t => {
       t.plan(1)
       await store.put({ id: 'a', v: 4 })
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 4, id: 'a' }, { v: 3, id: 'c' } ], 'range value' )
-    })
+    }))
 
-    t.test("add object 'b' to observed range", async t => {
+    t.test("add object 'b' to observed range", queued(async t => {
       t.plan(1)
       await store.put({ id: 'b', v: 5 })
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 4, id: 'a' }, { v: 5, id: 'b' }, { v: 3, id: 'c' } ], 'range value' )
-    })
+    }))
 
-    t.test("add object 'd' to observed range", async t => {
+    t.test("add object 'd' to observed range", queued(async t => {
       t.plan(1)
       await store.put({ id: 'd', v: 6 })
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 4, id: 'a' }, { v: 5, id: 'b' }, { v: 3, id: 'c' }, { v: 6, id: 'd' } ], 'range value' )
-    })
+    }))
 
-    t.test("unobserve range", async t => {
+    t.test("unobserve range", queued(async t => {
       t.plan(1)
       rangeObservable.unobserve(rangeObserver)
       t.pass('unobserved')
-    })
+    }))
+
   })
 
   t.test("observe range (a,d)", async t => {
@@ -93,45 +102,44 @@ test("store range observable", t => {
     let values = await getNextValue()
     t.deepEqual(values, [ { v: 5, id: 'b' }, { v: 3, id: 'c' } ], 'range value' )
 
-    t.test("remove object 'd' outside observed range", async t => {
+    t.test("remove object 'd' outside observed range", queued(async t => {
       t.plan(1)
       await store.delete('d')
       t.pass('deleted')
-    })
+    }))
 
-    t.test("remove object 'a' outside observed range", async t => {
+    t.test("remove object 'a' outside observed range", queued(async t => {
       t.plan(1)
       await store.delete('a')
       t.pass('deleted')
-    })
+    }))
 
-    t.test("add object 'ab' to observed range", async t => {
+    t.test("add object 'ab' to observed range", queued(async t => {
       t.plan(1)
       await store.put({ id: 'ab', v: 7 })
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 7, id: 'ab' }, { v: 5, id: 'b' }, { v: 3, id: 'c' } ], 'range value' )
-    })
+    }))
 
-    t.test("remove object 'ab' from observed range", async t => {
+    t.test("remove object 'ab' from observed range", queued(async t => {
       t.plan(1)
       await store.delete('ab')
       let values = await getNextValue()
       t.deepEqual(values, [ { v: 5, id: 'b' }, { v: 3, id: 'c' } ], 'range value' )
-    })
+    }))
 
-    t.test("unobserve range", async t => {
+    t.test("unobserve range", queued(async t => {
       t.plan(1)
       rangeObservable.unobserve(rangeObserver)
       t.pass('unobserved')
-    })
+    }))
+
   })
 
   t.test("close and remove database", async t => {
     t.plan(1)
     await level.close()
-    rimraf(dbPath, (err) => {
-      if(err) return t.fail(err)
-      t.pass('removed')
-    })
+    await rimraf(dbPath)
+    t.pass('removed')
   })
 })
