@@ -1,14 +1,14 @@
-const test = require('tape')
-const lmdb = require('node-lmdb')
-const rimraf = require("rimraf")
-const fs = require("fs")
+import test from 'tape'
+import lmdb from 'node-lmdb'
+import { rimraf } from 'rimraf'
+import fs from 'fs'
 
-const Store = require('../lib/Store.js')
+import Store from '../lib/Store.js'
 
-const dbPath = `./test.ro.db`
+const dbPath = `./test.co.db`
 rimraf.sync(dbPath)
 fs.mkdirSync(dbPath)
-const env = new lmdb.Env();
+const env = new lmdb.Env()
 env.open({
   // Path to the environment
   path: dbPath,
@@ -140,9 +140,7 @@ test("store count observable", t => {
     t.plan(1)
     dbi.close()
     env.close()
-    rimraf(dbPath, (err) => {
-      if(err) return t.fail(err)
-      t.pass('removed')
-    })
+    await rimraf(dbPath)
+    t.pass('removed')
   })
 })
