@@ -1,9 +1,9 @@
-const test = require('tape')
-const lmdb = require('node-lmdb')
-const rimraf = require("rimraf")
-const fs = require("fs")
+import test from 'tape'
+import lmdb from 'node-lmdb'
+import { rimraf } from 'rimraf'
+import fs from 'fs'
 
-const Store = require('../lib/Store.js')
+import Store from '../lib/Store.js'
 
 const dbPath = `./test.oo.db`
 rimraf.sync(dbPath)
@@ -74,9 +74,7 @@ test("store object observable", t => {
     t.plan(1)
     dbi.close()
     env.close()
-    rimraf(dbPath, (err) => {
-      if(err) return t.fail(err)
-      t.pass('removed')
-    })
+    await rimraf(dbPath)
+    t.pass('removed')
   })
 })
