@@ -1,6 +1,6 @@
-const test = require('tape')
+import test from 'tape'
 
-const Store = require('../lib/Store.js')
+import Store from '../lib/Store.js'
 
 test("store non-reactive properties", t => {
   t.plan(2)
